// generated/src/app/api-models/cov-prov-group-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CovProvGroupDetail {
  seqCovProvGrp: number;
  seqProvId: number;
  reimbMethod: string;
  effectiveDate: Date;
  termDate: Date;
  termReasn: string;
  userDefined1: string;
  userDefined2: string;
  userDefined3: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  seqDfltVendId: number;
  seqDfltVendAddress: number;
}